requirements.txt
--extra-index-url https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
torchvision==0.15.2+cpu
omegaconf
torchmetrics==0.10.3
fvcore
iopath
# xformers is currently not available for ARM, skip if necessary
# xformers==0.0.27
submitit
fastapi
uvicorn
requests
pillow
